Move all board props to selectors
