+++ src/components/Board/index.js
@@ -1,5 +1,4 @@
 import React, { useState, useLayoutEffect } from 'react';
-import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 import { useDispatch, useSelector } from 'react-redux';
 
@@ -8,6 +7,7 @@ import {
   getBottomPlayerIsRed,
   getNextMoveColor,
   getSelectedMove,
+  getLegalMoves,
 } from 'reducers';
 
 import * as logic from 'services/logic';
@@ -43,10 +43,11 @@ ${styles.MEDIA_LARGE} {
 
 const ANIMATION_DELAY = 150;
 
-const Board = ({ legalMoves }) => {
+const Board = () => {
   const dispatch = useDispatch();
-  const nextMoveColor = useSelector(state => getNextMoveColor(state));
   const bottomPlayerIsRed = useSelector(state => getBottomPlayerIsRed(state));
+  const legalMoves = useSelector(state => getLegalMoves(state));
+  const nextMoveColor = useSelector(state => getNextMoveColor(state));
   const selectedSlot = useSelector(state => state.selectedSlot);
   const {
     board, fromSlot: moveFromSlot, toSlot: moveToSlot,
@@ -148,10 +149,6 @@ const Board = ({ legalMoves }) => {
       {renderSquares()}
     </Wrapper>
   );
-};
-
-Board.propTypes = {
-  legalMoves: PropTypes.arrayOf(PropTypes.arrayOf(PropTypes.number)).isRequired,
 };
 
 export default Board;
